+++ ASCOfficeOdfFileW/source/Oox2OdfConverter/Converter.cpp
@@ -160,7 +160,7 @@ std::wstring EncodeBase64(const std::string & value)
 	char *pData = NULL;
 	std::wstring result;
 
-	NSFile::CBase64Converter::Encode((BYTE*)value.c_str(), value.length(), pData, nLength, NSBase64::B64_BASE64_FLAG_NOCRLF);
+	NSFile::CBase64Converter::Encode((BYTE*)value.c_str(), (int)value.length(), pData, nLength, NSBase64::B64_BASE64_FLAG_NOCRLF);
 	if (pData)
 	{
 		result = NSFile::CUtf8Converter::GetUnicodeStringFromUTF8((BYTE*)pData, nLength);
