+++ XlsxSerializerCom/Writer/BinaryCommonReader.h
@@ -50,11 +50,11 @@ namespace BinXlsxRW {
 	#define READ1_DEF(stLen, res, fReadFunction, arg) {\
 		long read1defCurPos = 0;\
 		long read1defstart_pos = m_oBufferedStream.GetPos();\
-		while(read1defCurPos < stLen)\
+		while(read1defCurPos < (long)stLen)\
 		{\
 			BYTE read1defType = m_oBufferedStream.GetUChar();\
 			ULONG read1defLength = m_oBufferedStream.GetULong();\
-			if (read1defLength + read1defCurPos > stLen)\
+			if (read1defLength + read1defCurPos > (ULONG)stLen)\
 			{\
 				m_oBufferedStream.Seek(read1defstart_pos + stLen);\
 				res = c_oSerConstants::ReadOk;\
